refactor: move derived-value evaluation into data-model
