+++ library/data-model/src/derivedFields/resolveParentField.ts
@@ -1,4 +1,27 @@
-import {DataEngine, getFieldToIdsMap} from '@faims3/data-model';
+/*
+ * Copyright 2026 Macquarie University
+ *
+ * Licensed under the Apache License Version 2.0 (the, "License");
+ * you may not use, this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing software
+ * distributed under the License is distributed on an "AS IS" BASIS
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND either express or implied.
+ * See, the License, for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Filename: resolveParentField.ts
+ * Description:
+ *   Resolves and formats a single parent field value for ParentFieldDisplay.
+ *   Lives in @faims3/data-model so the app and server-side evaluation share
+ *   one implementation.
+ */
+
+import {DataEngine} from '../databaseEngine/engine';
+import {getFieldToIdsMap} from '../uiSpecification';
 
 /** Outcome of resolving a parent field value for display. */
 export type ParentFieldResolution =
