+++ library/forms/lib/formModule/formManagers/parentReferences.test.ts
@@ -1,7 +1,7 @@
 import {describe, expect, it} from 'vitest';
 import {compileUiSpecConditionals, UiSpecModel} from '@faims3/data-model';
 import {recomputeDerivedFields} from './templatedFields';
-import {recomputeComputedFields} from './computedFields';
+import {recomputeComputedFields} from '@faims3/data-model';
 
 // Site parents Feature via a Child-relation Related Records field.
 const makeSpec = (): UiSpecModel => {
