+++ library/forms/lib/formModule/formManagers/resolveParentValues.test.ts
@@ -1,6 +1,6 @@
 import type {DataEngine, UiSpecModel} from '@faims3/data-model';
 import {describe, expect, it} from 'vitest';
-import {resolveParentValues} from './resolveParentValues';
+import {resolveParentValues} from '@faims3/data-model';
 
 const meta = {
   annotation: {include: false, label: 'annotation'},
